pyproject.toml
[tool.poetry]
name = "skypro-test"
version = "0.1.0"
description = ""
authors = ["samsegomof <wemagonof91@gmail.com>"]
readme = "README.md"
packages = [{include = "skypro_test"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.5"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
django-rest-framework = "^0.1.0"
django-filter = "^22.1"
django-admin-relation-links = "0.2.5"
drf-spectacular = "^0.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
